WebSockets/tests/submissions/Opera/binary/002.html

<!doctype html>
<title>WebSockets: Send/Receive blob, blob size greater than network array buffer</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':8007/echo');
  var data = "";
  var datasize = 100000;  
  ws.onopen = function(e) {
   ws.binaryType = "blob";
   data = new ArrayBuffer(datasize);
   ws.send(data);
  }
  ws.onmessage = function(e) {
    assertInstanceOf(e.data,Blob);
    assert_equals(e.data.size, datasize);
    end();
  }
  
});
</script>
